add tests for NumAsHex serialization/deserialization routines
