+++ crates/starknet-types-rpc/src/custom/syncing_status.rs
@@ -63,3 +63,26 @@ impl<'de> Deserialize<'de> for SyncingStatus {
         deserializer.deserialize_any(SyncingStatusVisitor)
     }
 }
+
+#[cfg(test)]
+#[test]
+fn syncing_status_from_false() {
+    let s = "false";
+    let syncing_status: SyncingStatus = serde_json::from_str(s).unwrap();
+    assert!(matches!(syncing_status, SyncingStatus::NotSyncing));
+}
+
+#[cfg(test)]
+#[test]
+fn syncing_status_to_false() {
+    let syncing_status = SyncingStatus::NotSyncing;
+    let s = serde_json::to_string(&syncing_status).unwrap();
+    assert_eq!(s, "false");
+}
+
+#[cfg(test)]
+#[test]
+fn syncing_status_from_true() {
+    let s = "true";
+    assert!(serde_json::from_str::<SyncingStatus>(s).is_err());
+}
