+++ crates/starknet-types-rpc/src/custom_serde/mod.rs
@@ -1,0 +1,5 @@
+//! Custom serialization and deserialization routines.
+
+mod num_as_hex;
+
+pub use self::num_as_hex::NumAsHex;
